refactor: 동작하지 않는 코드 제거 및 코드 정리
