+++ src/components/Title.jsx
@@ -1,11 +1,7 @@
+import { TITLE } from '../constants';
+
 const Title = () => {
-  return (
-    <h2>
-      국내 모든 임상시험 검색하고
-      <br />
-      온라인으로 참여하기
-    </h2>
-  );
+  return <h2>{TITLE}</h2>;
 };
 
 export default Title;
