+++ src/components/index.jsx
@@ -1,0 +1,6 @@
+import Container from './Container';
+import RecommendList from './RecommendList';
+import SearchBar from './SearchBar';
+import Title from './Title';
+
+export { Container, RecommendList, SearchBar, Title };
