+++ src/constants.js
@@ -1,1 +1,6 @@
 export const BASE_URL = 'https://api.clinicaltrialskorea.com';
+export const NO_DATA = '검색어 없음';
+export const RECOMMEND_KEYWORD = '추천 검색어';
+export const SEARCH_PLACEHOLDER = '질환명을 입력해주세요';
+export const SEARCH = '검색';
+export const TITLE = '국내 모든 임상시험 검색하고 온라인으로 참여하기';
