+++ src/styles/globals.css
@@ -1,23 +1,24 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  margin: 20px auto;
-  max-width: 500px;
+body {
+  background-color: #cae9ff;
 }
-
 form {
   display: flex;
   flex-direction: row;
   gap: 10px;
 }
-
 h2 {
   margin: 10px;
   font-size: 32px;
+  text-align: center;
 }
-
 input {
   width: 80%;
   height: 40px;
   padding-left: 10px;
+}
+.container {
+  display: flex;
+  flex-direction: column;
+  margin: 20px auto;
+  max-width: 450px;
 }
